+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev-master
 
+- Adapt minor changes of clue/graph dependency
 - Fix for collecting dependencies of null filltered ADTs
 - Fix in realpath usage for message about write target
 - Add Location awareness to Verteces and Edges